refactor: externalize template and modularize filtering
